overturemaps_wfsserver/requirements.txt
Django>=3.0,<4.0
psycopg2-binary
gdal==3.4.1
django-gisserver==1.4.0
geojson
geopandas
shapely
overturemaps==0.5.0
django-cors-headers
